client: move http.go into client.go
